program: Add canonical PDA for stakes per validator
#### Problem

The mSOL program allows for many different stake accounts delegated to a
single validator, which is very flexible, but can make accounting more
difficult, since there's no one place for a particular validator's
stake.

#### Summary of changes

Following a model similar to spl-stake-pool, enforce that all stake
eventually converges to a PDA based on the pool and validator pubkeys.

* Introduce a new instruction to create the PDA based on a stake account
  delegated to a particular validator, by splitting everything into the
  target account
* Abort merge-stake if the destination is not the PDA

After this change, before attempting a merge, bots must check if the
canonical PDA exists for the validator, and add a `CreateCanonicalStake`
instruction. And then all merges must go into the PDA.

Other than that, the changes are minimal.